DRY building a test DB, and remove unused stuff
  - There should only be one way to build a test db
  - `DatabaseManager/main.py` was previously moved to a management command
  - Remove scripts which no longer work and are unlikely to be used
