+++ Pipfile
@@ -41,9 +41,6 @@ whitenoise = "*"
 foma = {subdirectory = "foma/python", git = "https://github.com/andrewdotn/foma"}
 
 [scripts]
-# look stuff up
-analyze = "hfst-optimized-lookup CreeDictionary/res/fst/crk-descriptive-analyzer.hfstol"
-generate = "hfst-optimized-lookup CreeDictionary/res/fst/crk-normative-generator.hfstol"
 # unit tests
 test = "pytest --mypy CreeDictionary/"
 # run dev server
@@ -51,21 +48,10 @@ dev = "python CreeDictionary/manage.py runserver"
 runserver_for_tests = "env USE_TEST_DB=true python CreeDictionary/manage.py runserver"
 # format
 format = "black CreeDictionary"
-# `import-xml` removes the database and builds a new one from the start,
-# it shouldn't be used because we are waiting for a well structured new database
-# so that we can move to a more continuous workflow where manual database modification can be persisted
-# example usage: pipenv run import-xml CreeDictionary/res/dictionaries/ --multi-process 2
-import-xml = "sh ./libexec/import.sh"
 # building the test db takes about half a minute
-build-test-db = "./libexec/build_test_db.sh"
-# warning: do not use these.
-# eventually we will scrap them and use a migration based work flow to preserve database changes AFTER a new
-# , well structured database is ready.
-push-db = 'rsync --progress CreeDictionary/db.sqlite3 $SAPIR_USER@sapir.artsrn.ualberta.ca:/opt/cree-intelligent-dictionary/CreeDictionary/db.sqlite3'
-pull-db = 'rsync --progress $SAPIR_USER@sapir.artsrn.ualberta.ca:/opt/cree-intelligent-dictionary/CreeDictionary/db.sqlite3 CreeDictionary/db.sqlite3'
+ensuretestdb = "env USE_TEST_DB=true python CreeDictionary/manage.py ensuretestdb"
 # Utitlities
 collect-static = "python CreeDictionary/manage.py collectstatic --no-input"
-gunicorn-locally = "gunicorn CreeDictionary.wsgi"
 
 [requires]
 python_version = "3.9"
